+++ CHANGELOG.md
@@ -15,6 +15,8 @@ duolingo plus.
 ### Fixed
 - Class name values for the top bar and its navigation buttons.
 - Class name values for language change, crowns info and streak info popups.
+- Handling of changing between mobile and desktop layout while crown and XP
+popups are displayed.
 
 [v1.3.9] - 2020-05-03
 -----------------